NewChat/requirement.txt
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
decorator==5.2.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
executing==2.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipdb==0.13.13
ipython==8.34.0
jedi==0.19.2
matplotlib-inline==0.1.7
msgpack==1.1.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
redis==5.2.1
service-identity==24.2.0
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
wcwidth==0.2.13
zope.interface==7.2
